% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPIT.R
\name{davies_exact}
\alias{davies_exact}
\title{Runs the Davies portion of the hypothesis testing}
\usage{
davies_exact(vc.mod, X)
}
\arguments{
\item{vc.mod}{The return object of the MAPIT function.}

\item{X}{A vector containing names.}
}
\value{
The p-values for association.
}
\description{
Runs the Davies portion of the hypothesis testing
}
