% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_genotype_matrix.R
\docType{data}
\name{random_genotype_matrix}
\alias{random_genotype_matrix}
\title{WTCCC Genotype Matrix.}
\format{
A data frame with 5747 rows and 2938 variables:
\describe{
  \item{sample_name}{name of the sample}
  \item{SNPs}{O for homozygous in major allele, 1 for heterozygous, 2 for homozygous in minor allele}
}
}
\source{
\url{https://www.wtccc.org.uk/}
}
\usage{
random_genotype_matrix
}
\description{
A dataset containing 5747 samples with 2938 SNPs.
}
\keyword{datasets}
