% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{simulate
This function will run a version of the MArginal ePIstasis Test (MAPIT) under the following model variations:
(1) Standard Model: y = m+g+e where m ~ MVN(0,omega^2K), g ~ MVN(0,sigma^2G), e ~ MVN(0,tau^2M).
Recall from Crawford et al. (2017) that m is the combined additive effects from all other variants,
and effectively represents the additive effect of the kth variant under the polygenic background
of all other variants; K is the genetic relatedness matrix computed using
genotypes from all variants other than the kth; g is the summation of all pairwise interaction
effects between the kth variant and all other variants; G represents a relatedness matrix
computed based on pairwise interaction terms between the kth variant and all other variants. Here,
we also denote D = diag(x_k) to be an n × n diagonal matrix with the genotype vector x_k as its
diagonal elements. It is important to note that both K and G change with every new marker k that is
considered. Lastly; M is a variant specific projection matrix onto both the null space of the intercept
and the corresponding genotypic vector x_k.
(2) Standard + Covariate Model: y = Wa+m+g+e where W is a matrix of covariates with effect sizes a.
(3) Standard + Common Environment Model: y = m+g+c+e where c ~ MVN(0,eta^2C) controls for extra
environmental effects and population structure with covariance matrix C.
(4) Standard + Covariate + Common Environment Model: y = Wa+m+g+c+e
This function will consider the following three hypothesis testing strategies which are featured in Crawford et al. (2017):
(1) The Normal or Z test
(2) Davies Method
(3) Hybrid Method (Z test + Davies Method)}
\usage{
simulate(
  nsnps,
  npepi,
  ntepi,
  H2 = 0.6,
  d = 2,
  rho = 0.8,
  seed = 67132,
  data = "random_genotype_matrix",
  outfile = NULL
)
}
\arguments{
\item{nsnps}{Number of causal SNPs}

\item{npepi}{Number of pleiotropic epistatic SNPs. npepi+ntepi must be less than nsnps.}

\item{ntepi}{Number of trait specific epistatic SNPs. npepi+ntepi must be less than nsnps.}

\item{H2}{Broad-sense heritability. Default 0.6.}

\item{d}{Number of phenotypes. Default 2.}

\item{rho}{Proportion of heritability explained by additivity. Default 0.8.}

\item{seed}{Random seed for simulation. Default 67132.}

\item{data}{Data frame from which to draw SNPs. Default random_genotype_matrix.}

\item{outfile}{is the filename for the output file.}
}
\value{
A list of P values and PVEs
}
\description{
simulate
This function will run a version of the MArginal ePIstasis Test (MAPIT) under the following model variations:
(1) Standard Model: y = m+g+e where m ~ MVN(0,omega^2K), g ~ MVN(0,sigma^2G), e ~ MVN(0,tau^2M).
Recall from Crawford et al. (2017) that m is the combined additive effects from all other variants,
and effectively represents the additive effect of the kth variant under the polygenic background
of all other variants; K is the genetic relatedness matrix computed using
genotypes from all variants other than the kth; g is the summation of all pairwise interaction
effects between the kth variant and all other variants; G represents a relatedness matrix
computed based on pairwise interaction terms between the kth variant and all other variants. Here,
we also denote D = diag(x_k) to be an n × n diagonal matrix with the genotype vector x_k as its
diagonal elements. It is important to note that both K and G change with every new marker k that is
considered. Lastly; M is a variant specific projection matrix onto both the null space of the intercept
and the corresponding genotypic vector x_k.
(2) Standard + Covariate Model: y = Wa+m+g+e where W is a matrix of covariates with effect sizes a.
(3) Standard + Common Environment Model: y = m+g+c+e where c ~ MVN(0,eta^2C) controls for extra
environmental effects and population structure with covariance matrix C.
(4) Standard + Covariate + Common Environment Model: y = Wa+m+g+c+e
This function will consider the following three hypothesis testing strategies which are featured in Crawford et al. (2017):
(1) The Normal or Z test
(2) Davies Method
(3) Hybrid Method (Z test + Davies Method)
}
