% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_epistasis_data.R
\docType{data}
\name{simulated_epistasis_data}
\alias{simulated_epistasis_data}
\title{Genotype and phenotype data with epistasis.}
\format{
A list with simulated data and simulation parameters:
\describe{
  \item{numer_snp}{Number of SNPs for each genotype.}
  \item{number_samples}{Number of genotype samples.}
  \item{number_causal_snp}{Number of causal SNPs.}
  \item{pve}{Phenotypic variance explained/broad-sense heritability (H^2).}
  \item{rho}{Portion of H^2 that is contributed to by the marginal (additive) effects.}
  \item{phenotype}{The simulated phenotype. Vector size (number_samples) x 1.}
  \item{genotype}{The simulated genotype; scaled and centered. Matrix size (number_samples) x (number_snps).}
  \item{linear_effects}{Simulated linear effect sizes. Vector size (number_causal_snp) x 1.}
  \item{interaction_effects}{Simulated epistatic effect sizes. Vector size (length(epistatic_snps)) x 1.}
  \item{epistatic_snps}{Matrix entries take values 0,1,2. 0 for homozygous in major allele, 1 for heterozygous, 2 for homozygous in minor allele.}
  \item{additive_snps}{Matrix entries take values 0,1,2. 0 for homozygous in major allele, 1 for heterozygous, 2 for homozygous in minor allele.}
}
}
\source{
data-raw/simulate_epistasis.R
}
\usage{
simulated_epistasis_data
}
\description{
A simulated dataset that has epistatic interactions.
}
\keyword{datasets}
