% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pvalues.R
\name{harmonic_combined}
\alias{harmonic_combined}
\title{Harmonic mean p combine method on MvMAPIT return}
\usage{
harmonic_combined(pvalues)
}
\arguments{
\item{pvalues}{Tibble with p-values from MvMAPIT function call. Grouping is
based on the column named "id"}
}
\value{
A Tibble with the combined p-values.
}
\description{
This function takes in the p-values tibble that MvMAPIT returned. It then
computes the combined p-values grouped by variant id.
}
\examples{
set.seed(837)
p <- 200
n <- 100
d <- 2
X <- matrix(
    runif(p * n),
    ncol = p
)
Y <- matrix(
    runif(d * n),
    ncol = d
)
mapit <- mvmapit(
    t(X),
    t(Y),
    test = "normal", cores = 1, logLevel = "INFO"
)
harmonic <- harmonic_combined(mapit$pvalues)
}
