% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_phenotypes}
\alias{simulate_phenotypes}
\title{Simulate phenotye data}
\usage{
simulate_phenotypes(
  genotype_matrix,
  causal_fraction = 0.2,
  epistatic_fraction = 0.5,
  pleiotropic_fraction = 0.5,
  H2 = 0.6,
  d = 2,
  rho = 0.8,
  marginal_correlation = 0.3,
  epistatic_correlation = 0.3,
  seed = 67132
)
}
\arguments{
\item{genotype_matrix}{Genotype matrix with samples as rows, and SNPs as columns.}

\item{causal_fraction}{Fraction of the SNPs that are causal.}

\item{epistatic_fraction}{Fraction of the causal SNPs with single trait epistatic effects.}

\item{pleiotropic_fraction}{Fraction of SNPs with pleiotropic effects.}

\item{H2}{Broad-sense heritability.}

\item{d}{Number of phenotypes.}

\item{rho}{Proportion of heritability explained by additivity.}

\item{marginal_correlation}{Correlation between the additive effects of the phenotype.}

\item{epistatic_correlation}{Correlation between the epistatic effects of the phenotype.}

\item{seed}{Random seed for simulation.}
}
\value{
A list object containing the phenotype data, the genotype data, as well as the causal SNPs and summary statistics.
}
\description{
\code{simulate_phenotype} simulates phenotype data from a genotype matrix.
}
\details{
This function takes a genotype matrix and simulates phenotype data under the following model:

beta_i ~ MN(0, V_i, I), i in \{ additive, epistatic, residual\}

The effect sizes follow a matrix normal distribution with no correlation between the samples but covariance matrix between phenotypes.
}
